New Main Scene
* Moved code to connect scene planks into its own script
* Made a new house scene
* Made new main scene with player and house
