+++ Building/BuildingTest.tscn
@@ -1,9 +1,8 @@
-[gd_scene load_steps=6 format=2]
+[gd_scene load_steps=5 format=2]
 
 [ext_resource path="res://Building/Plank.tscn" type="PackedScene" id=1]
 [ext_resource path="res://Building/BuildingTest.cs" type="Script" id=2]
-[ext_resource path="res://Building/Plank_Big.tscn" type="PackedScene" id=3]
-[ext_resource path="res://Building/Plank_Small.tscn" type="PackedScene" id=4]
+[ext_resource path="res://Building/House.tscn" type="PackedScene" id=3]
 
 [sub_resource type="RectangleShape2D" id=1]
 extents = Vector2( 529, 8 )
@@ -32,63 +31,5 @@ position = Vector2( 0, 38 )
 position = Vector2( -32, -22 )
 rotation = -1.0472
 
-[node name="House" type="Node2D" parent="."]
-position = Vector2( 359, 510 )
-
-[node name="BigPlank(RigidBody2D)" parent="House" instance=ExtResource( 3 )]
-position = Vector2( 208, -70 )
-
-[node name="BigPlank(RigidBody2D)2" parent="House" instance=ExtResource( 3 )]
-position = Vector2( 209, -364 )
-
-[node name="BigPlank(RigidBody2D)5" parent="House" instance=ExtResource( 3 )]
-position = Vector2( 108, -377 )
-rotation = -0.785397
-
-[node name="BigPlank(RigidBody2D)6" parent="House" instance=ExtResource( 3 )]
-position = Vector2( 311, -377 )
-rotation = 0.785397
-
-[node name="BigPlank(RigidBody2D)3" parent="House" instance=ExtResource( 3 )]
-position = Vector2( 62, -215 )
-rotation = 1.57079
-
-[node name="BigPlank(RigidBody2D)4" parent="House" instance=ExtResource( 3 )]
-position = Vector2( 356, -216 )
-rotation = 1.57079
-
-[node name="Plank(RigidBody2D)8" parent="House" instance=ExtResource( 1 )]
-position = Vector2( 78, 3 )
-rotation = -1.57096
-
-[node name="Plank(RigidBody2D)16" parent="House" instance=ExtResource( 1 )]
-position = Vector2( 340, 3 )
-rotation = -1.57096
-
-[node name="Plank(RigidBody2D)17" parent="House" instance=ExtResource( 1 )]
-position = Vector2( 209, 2 )
-rotation = -1.57096
-
-[node name="SmallPlank(RigidBody2D)" parent="House" instance=ExtResource( 4 )]
-position = Vector2( 194, -42 )
-rotation = 1.0472
-
-[node name="SmallPlank(RigidBody2D)3" parent="House" instance=ExtResource( 4 )]
-position = Vector2( 224, -41 )
-rotation = -1.0472
-
-[node name="SmallPlank(RigidBody2D)4" parent="House" instance=ExtResource( 4 )]
-position = Vector2( 99, -52 )
-rotation = -0.785397
-
-[node name="SmallPlank(RigidBody2D)5" parent="House" instance=ExtResource( 4 )]
-position = Vector2( 77, -97 )
-rotation = 0.785398
-
-[node name="SmallPlank(RigidBody2D)8" parent="House" instance=ExtResource( 4 )]
-position = Vector2( 340, -93 )
-rotation = -0.785398
-
-[node name="SmallPlank(RigidBody2D)2" parent="House" instance=ExtResource( 4 )]
-position = Vector2( 317, -48 )
-rotation = 0.785397
+[node name="House" parent="." instance=ExtResource( 3 )]
+position = Vector2( 618, 591 )
